### 2016-07-01 #34 @whedon
Submission: Osprey: Hyperparameter Optimization for Machine Learning
**Submitting author:** @cxhernandez (<a href="http://orcid.org/0000-0002-8146-5904">Carlos Xavier Hernández</a>)
**Repository:** <a href="https://github.com/msmbuilder/osprey" target ="_blank">https://github.com/msmbuilder/osprey</a>
**Version:** v1.1.0
