### 2016-07-01 #34 @arfon
Hi @cxhernandez. Thanks for submitting this package to JOSS. Unfortunately it doesn't look like there's a `paper.md` file in this repository.

Could you please take a look at the [submission guidelines](http://joss.theoj.org/about#author_guidelines) and make sure there's a paper present in the repository. We can then move forward in the review process.
