+++ devtools/conda-recipe/meta.yaml
@@ -1,6 +1,6 @@
 package:
   name: osprey
-  version: "1.0.0"
+  version: "1.1.0dev0"
 
 source:
   path: ../../
