+++ paper/paper.bib
@@ -41,6 +41,22 @@ url="http://dx.doi.org/10.1023/A:1008306431147"
     Booktitle = {Proceedings of the 30th International Conference on Machine Learning (ICML-13)},
     Pages = {115-123}
    }
+ @misc{msmbuilder,
+  title={msmbuilder: MSMBuilder 3.4},
+  url={http://dx.doi.org/10.5281/zenodo.48545},
+  DOI={10.5281/zenodo.48545},
+  publisher={Zenodo},
+  author={Robert T. McGibbon and
+          Matthew Harrigan and
+          Bharath Ramsundar and
+          Kyle Beauchamp and
+          Muneeb Sultan and
+          Christian Schwantes and
+          Carlos Hernández and
+          Peter Eastman and
+          Brooke Husic and
+          Patrick Riley and et al.},
+  year={2016}}
 @misc{osprey_archive,
   author       = {Robert T. McGibbon and
                   Carlos Hernández and
