+++ paper/paper.bib
@@ -50,21 +50,21 @@ url="http://dx.doi.org/10.1023/A:1008306431147"
           Matthew Harrigan and
           Bharath Ramsundar and
           Kyle Beauchamp and
-          Muneeb Sultan and
+          Mohammad M. Sultan and
           Christian Schwantes and
-          Carlos Hernández and
+          Carlos X. Hernández and
           Peter Eastman and
-          Brooke Husic and
+          Brooke E. Husic and
           Patrick Riley and et al.},
   year={2016}}
 @misc{osprey_archive,
   author       = {Robert T. McGibbon and
-                  Carlos Hernández and
-                  Matthew Harrigan and
+                  Carlos X. Hernández and
+                  Matthew P. Harrigan and
                   Steven Kearnes and
-                  Muneeb Sultan and
+                  Mohammad M. Sultan and
                   Stanislaw Jastrzebski and
-                  Brooke Husic and
+                  Brooke E. Husic and
                   Vijay Pande},
   title        = {osprey: Osprey 1.0.0},
   month        = jun,
