### 2016-07-01 #34 @cxhernandez
Hey @arfon,

Sorry about that! I've added the paper to the repository now and it can be found [here](https://github.com/msmbuilder/osprey/tree/master/paper).

Thanks for your consideration!
