### 2016-07-29 #34 @rasbt
@labarba @arfon Thanks for asking me to review the software;  it's a really interesting package that may turn out to be useful in some research projects. Being new to JOSS, I just read through the guidelines and posted a summary below. It would be nice if you could give me "permission" to edit the review sheet at the top.

@cxhernandez
Thanks for your submission. Although it seems like I listed a fair amount of points below, I must say that I find osprey really appealing! It would be great though if you could address these points, and I am happy to discuss them further :). I think the most important thing would be to fix the unit tests so that they work with current scikit-learn versions (or set `scikit-learn=0.16` (?) as a hard requirement) so that I can look into the package further and try out some of the examples (and take it to a whirl on some real world data).

## Software

The Read-the-Docs style documentation at http://msmbuilder.org/osprey/1.0.0/ looks nice to me. It would be nice to link back to the GitHub repository though. Also, I noticed that this is the documentation for version 1.0 and in "Installation" it says that the release version can be installed via  `pip install osprey`, which is v0.4.

- [ ] tests

There's currently only a 63% coverage; I am not sure what the JOSS requirements are, but I would recommend extending the unit tests to at least 90%.
Also, I am unable to run the unit tests on my system

>>> from sklearn.utils import check_arrays
ImportError: cannot import name 'check_arrays'
```

I think this is a left-over from a very old scikit-learn version (< 0.16); the `check_arrays` function has been removed a long time ago (there's a different `check_array` now).

Also, I think the unitests are currently not running on Travis CI?

- [ ] License

There is a Apache 2.0 license template in the repo, however, it needs to be filled out with the necessary information to make it a valid license. E.g., it currently says:

> [...] Copyright [yyyy] [name of copyright owner]

Also, the "license badge" in the README links to the PyPI package not to the license file itself.

I would suggest to mention the license (and add a link) on the documentation website as well.

## The JOSS paper

The JOSS paper (the PDF associated with this submission) should only include:

-  [x] A list of the authors of the software

- [ ] Author affiliations

Author affilations are listed, however, I suggest removing the redundancy, i.e., "Stanford University" is currently listed 7 times. It would be sufficient to list it one time and change the author "indices" accordingly.

- [ ] A short summary describing the high-level functionality of the software

The authors provide a nicely written summary explaining the purpose, functionality, and underlying concepts in a brief, yet informative way including references to relevant literature. Great!

- [ ] A list of key references including a link to the software archive

The "Software Repository:" and "Software Archive:" rows are currently blank and the respective links need to be provided.

- [x] A high-level overview of this documentation should be included in a README file (or equivalent).

Nice README file!

## A statement of need

- [x] The authors should clearly state what problems the software is designed to solve and who the target audience is.

The purpose of the software is summarized nicely

## Installation instructions

- [x] There should be a clearly-stated list of dependencies. Ideally these should be handled with an automated package management solution.

The authors provide a list of dependencies in the body of the README file. I would suggest to add these to an additional `requirements.txt` file so that they can be installed via pip in one swoop.

Only NumPy is listed as dependency, not SciPy -- the latter is marked as "optional". However, both are required by scikit-learn, which is listed as a dependency as well. I suggest either listing both SciPy and NumPy as hard dependencies or neither of them (since they are sklearn dependencies).

Also, it would be nice to mention the version numbers of the packages that are required (e.g., right now, unit tests don't run with scikit-learn > 0.16)

## Example usage

- [x] The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

The examples are brief but clear and sufficient :).

## API documentation

- [x] Reviewers should check that the software API is documented to a suitable level.

Only very few functions have proper docstrings, but I want to give it an "okay" here since the user mostly interacts with osprey via a config file, which is then executed via `osprey worker config.yaml`; more docstrings would be nice for contributors though.

- [x] Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

Tests exist but currently don't run on Travis (?) and they don't run with current package version of e.g., scikit-learn like mentioned earlier.

There should be clear guidelines for third-parties wishing to:

- [ ] Contribute to the software

I couldn't find any guidelines.

- [ ] Report issues or problems with the software

Also, I didn't see anything here. Putting a link to GitHub issue tracker would be sufficient to address this though.

- [ ] Seek support

Didn't find any information on this.

> Include here some examples of well-documented software for people to review.

E.g., scikit-learn, Keras, or TensorFlow

- [ ] Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.

See unit test issues I mentioned above.
