### 2016-07-29 msmbuilder/osprey#160 @cxhernandez
Update PyPi version
Still on v0.4 in https://pypi.python.org/pypi/osprey/. @rmcgibbo could you update this when you get a chance or give me maintainer access?
