### 2016-07-29 #34 @cxhernandez
Thanks @rasbt for reviewing our work! I've tried to address everything point-by-point:

## Software

### Documentation

> It would be nice to link back to the GitHub repository though.

Thanks for the suggestion! I'll add a link to the GitHub repo on the homepage.

> Also, I noticed that this is the documentation for version 1.0 and in "Installation" it says that the release version can be installed via pip install osprey, which is v0.4.

Good catch! We have [not updated the PyPI package yet](https://github.com/msmbuilder/osprey/issues/160). In the meantime, the recommended way to [install the latest version](https://github.com/msmbuilder/osprey#installation) is via our conda repo (`conda install -c omnia osprey`). We'll update our PyPI package ASAP

### Tests

> There's currently only a 63% coverage; I am not sure what the JOSS requirements are, but I would recommend extending the unit tests to at least 90%.

We're working on improving our test coverage; however, if you look closely a good deal of the "untested" code (as reported by [coveralls](https://coveralls.io/github/msmbuilder/osprey?branch=master)) is part of our CLI codebase, for which there are, in fact, [unit tests](https://github.com/msmbuilder/osprey/blob/master/osprey/tests/test_cli_worker_and_dump.py).

> Also, I am unable to run the unit tests on my system....

This is probably from running v0.4, this import is no longer in the [latest version](https://github.com/msmbuilder/osprey/search?utf8=%E2%9C%93&q=check_arrays) (v1.0).

> Also, I think the unitests are currently not running on Travis CI?

Unit tests have indeed been running on [Travis CI](https://travis-ci.org/msmbuilder/osprey).

### License

> Also, the "license badge" in the README links to the PyPI package not to the license file itself.
>
> I would suggest to mention the license (and add a link) on the documentation website as well.

I'll make these fixes ASAP and add a link on the homepage as well.

## Paper

>  I suggest removing the redundancy, i.e., "Stanford University" is currently listed 7 times

The affiliations listing is taken directly from the [JOSS LaTeX template](https://github.com/openjournals/whedon/blob/master/resources/latex.template). However, I completely agree that this is redundant information and will update the affiliations accordingly.

> The "Software Repository:" and "Software Archive:" rows are currently blank and the respective links need to be provided.

I'll update the "Software Repository" and "Software Archive" ASAP

## Installation instructions

> I would suggest to add these to an additional requirements.txt file so that they can be installed via pip in one swoop.

I'll add this ASAP

> Only NumPy is listed as dependency, not SciPy -- the latter is marked as "optional".

Good point! I can update the README to list SciPy as a full dependency.

> Also, it would be nice to mention the version numbers of the packages that are required (e.g., right now, unit tests don't run with scikit-learn > 0.16)

Osprey v1.0 should work with latest versions of its dependencies (unless specified). This is also probably a result of using v0.4.

## API Documentation

> more docstrings would be nice for contributors though

Agreed. The documentation has been tailored more for users at the moment, but we plan to include sections on creating/adding custom classes for `dataset_loaders`, etc.

## Tests

> Tests exist but currently don't run on Travis (?) and they don't run with current package version of e.g., scikit-learn like mentioned earlier.

They should run on [Travis CI](https://travis-ci.org/msmbuilder/osprey). Not sure what evidence there is to the contrary.

## Community guidelines

This is something we are admittedly lacking; however, I can add a section about contributing to our README and homepage. All development occurs on GitHub, so we rely on the issue tracker for troubleshooting and support, as well as accepting software contributions via pull request.
