### 2016-07-29 msmbuilder/osprey#160 @rmcgibbo
Thanks for the poke.

-Robert

On Fri, Jul 29, 2016 at 4:52 PM, Carlos Hernández <notifications@github.com>
wrote:

> Assigned #160 <https://github.com/msmbuilder/osprey/issues/160> to
> @rmcgibbo <https://github.com/rmcgibbo>.
>
