### 2016-07-30 #34 @rasbt
Wow, that was quick :) Thanks for addressing and commenting on these points so thoroughly!

> Osprey v1.0 should work with latest versions of its dependencies (unless specified). This is also probably a result of using v0.4.

> Good catch! We have not updated the PyPI package yet. In the meantime, the recommended way to install the latest version is via our conda repo (conda install -c omnia osprey). We'll update our PyPI package ASAP

Hm, I had an older version indeed. However, I did install the conda version; I also suggest updating it to 1.0.

```bash
~$ conda install -c omnia osprey
Fetching package metadata .........
Solving package specifications: ..........

Package plan for installation in environment /Users/Sebastian/miniconda3:

The following NEW packages will be INSTALLED:

    osprey: 0.4-py35_0 omnia

Proceed ([y]/n)?
```

> We're working on improving our test coverage; however, if you look closely, a good deal of the "untested" code (as reported by coveralls) is part of our CLI codebase, for which there are, in fact, unit tests.

Sounds good! However, I am bit confused now, if there are unit tests, why is the coverage 0.0 for these then? Also, why did you prepend underscores (`_`) in some cases? E.g.,

```
def _test_dump_1():
    out = subprocess.check_output(
        [OSPREY_BIN, 'dump', 'config.yaml', '-o', 'json'])
    if sys.version_info >= (3, 0):
        out = out.decode()
    json.loads(out)
```

Is it due to some complications with travis and subprocess?

> Unit tests have indeed been running on Travis CI.

Ah, sorry about that ... I see it now. Have only looked at the travis.yaml and somehow assumed that it wasn't running nose at all.

> The affiliations listing is taken directly from the JOSS LaTeX template. However, I completely agree that this is redundant information and will update the affiliations accordingly.

Now that you mentioned it, I noticed the same problem with other papers as well. @arfon Maybe it'd be worthwhile updating the template to make it easier on the authors in future submissions?

> Osprey v1.0 should work with latest versions of its dependencies (unless specified). This is also probably a result of using v0.4.

Nice. One more thing, though :P Could you please add the minimum version numbers to the README (and maybe the requirements.txt file for people who install it via pip) for the dependencies? E.g., something like

```
scipy>=0.17
numpy>=1.10.4
scikit-learn>=0.17
...
```

> This is something we are admittedly lacking; however, I can add a section about contributing to our README and homepage. All development occurs on GitHub, so we rely on the issue tracker for troubleshooting and support, as well as accepting software contributions via pull request.

That's fine and probably the best solution! I would just add a sentence to the README (and website) like

"In case you encounter any issues with this package, please consider submitting a ticket to the [GitHub Issue Tracker](https://github.com/msmbuilder/osprey/issues). We also welcome feature requests, and please don't hesitate to submit pull requests for bug fixes and improvements."

Or maybe consider setting up a "Contributing" section with some more info on "how-to contribute" info, which would be even nicer :)

Let me know when you updated the conda installer and PyPI version; I am happy to take the 1.0 to a test drive ;)
