### 2016-07-30 #34 @arfon
> @arfon Maybe it'd be worthwhile updating the template to make it easier on the authors in future submissions?

@rasbt - do you know if there's a way to handle this in LaTeX automatically? I agree this is less-than-ideal but I'm not exactly sure how to fix this.
