### 2016-07-30 #34 @cxhernandez
> Let me know when you updated the conda installer and PyPI version; I am happy to take the 1.0 to a test drive ;)

@rasbt: It should now be available through conda and pip. Sorry about that!
