### 2016-08-01 #34 @rasbt
No worries, @cxhernandez thanks for updating it, it seems to work fine now and I will have a look at the remaining points in the review (I will go through the previous points again and check off the list items)

- [ ] One more thing that I'd recommend you to address:

In the README.md (and on PyPI), it says

---

If you have an Anaconda Python distribution, installation is as easy as:

$ conda install -c omnia osprey
You can also install with pip:

$ pip install git+git://github.com/pandegroup/osprey.git
Alternatively, you can install directly from this GitHub repo:

---

I would change the line

    pip install git+git://github.com/pandegroup/osprey.git

to

    pip install osprey

so that people install the latest stable version by default, not the developer version. I would mention the "developer version" via `pip install git+git://github.com/pandegroup/osprey.git` on a separate line then.
