### 2016-08-01 #34 @rasbt
@arfon

I usually do this by assigning the same author index to 2 authors if they have the same affiliation:

```
\documentclass{article}

\usepackage{authblk}

\title{My paper}

\author[1]{Firstname1 Lastname1}
\author[2]{Firstname2 Lastname2}
\author[1]{Firstname3 Lastname3}

\affil[1]{Department of Math, University A}
\affil[2]{Department of Math, University B}

\begin{document}

\maketitle

\end{document}
```

![screen shot 2016-08-01 at 3 59 09 pm](https://cloud.githubusercontent.com/assets/5618407/17306999/e84b5b4e-5800-11e6-90ec-ba8f378c7679.png)

I only looked at the template briefly, but I think that's going to be tricky, you are right. I am not sure how to do this directly in the template, and the only spontaneous, ugly workaround I could think of would be to write a temporary tex file and modify this with a custom script

1. pandoc: .md -> .tex
2. script: fix numbers for duplicate author affiliations
3. pandoc: .tex -> .pdf
