### 2016-08-01 #34 @mpharrigan
You could use a style like `revtex` to automatically condense duplicate affiliations, or look at their `.sty` file to figure out what trickery they're using to do so
