### 2016-08-01 #34 @rasbt
@cxhernandez

please add

- [ ] cython
- [ ] mdtraj
- [ ] msmbuilder
- [ ] hdf5
- [ ] numexpr
- [ ] tables
- [ ] matplotlib
- [ ] numpydoc
- [ ] bokeh
- [ ] pandas

to the list of requirements. I wasn't able to run the unit tests without them.
