### 2016-08-01 #34 @rasbt
@cxhernandez

Seems to work well so far :). However, I am currently not able to reproduce the results, how can I set the random state in the configuration file? I couldn't find a note in the documentation.
