### 2016-08-01 #34 @cxhernandez
@rasbt: We don't currently have the option to set the random state in the configuration file, but that's a great suggestion and I'll look into implementing it!
