### 2016-08-01 msmbuilder/osprey#162 @cxhernandez
Set random state in config
see @rasbt's suggestion: https://github.com/openjournals/joss-reviews/issues/34#issuecomment-236711338
