### 2016-08-01 #34 @rasbt
@cxhernandez

I am currently having some other issues with alternative search strategies. E.g., the following one works fine:

```
estimator:
  entry_point: sklearn.svm.SVC

strategy:
  name: random

search_space:
  C:
    min: 0.1
    max: 10
    type: float

  gamma:
    min: 1e-5
    max: 1
    warp: log
    type: float

cv: 5

dataset_loader:
  name: sklearn_dataset
  params:
    method: load_iris

trials:
    uri: sqlite:////Users/Sebastian/Desktop/osprey-trials.db
```

However, when I switch from

```
strategy:
  name: random
```

to

```
strategy:
  name: grid
```

I get

```
Loading trials database: sqlite:////Users/Sebastian/Desktop/osprey-trials.db...
History contains: 10 trials
Choosing next hyperparameters with grid...
Error: GridSearchStrategy is defined only for all-enum search space
```

When I choose

```
strategy:
  name: gp
```

```
    return [var.point_to_gp(point_dict[var.name]) for var in self]
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/search_space.py", line 112, in <listcomp>
    return [var.point_to_gp(point_dict[var.name]) for var in self]
KeyError: 'C'
```
