### 2016-08-01 9b2d2b9
add examples; fix jump var
