+++ examples/config-2.yaml
@@ -2,7 +2,9 @@ estimator:
   entry_point: sklearn.svm.SVC
 
 strategy:
-  name: random
+  name: gp
+  params:
+    seeds: 5
 
 search_space:
   C:
