+++ examples/config-3.yaml
@@ -2,19 +2,23 @@ estimator:
   entry_point: sklearn.svm.SVC
 
 strategy:
-  name: random
+  name: grid
 
 search_space:
   C:
     min: 0.1
     max: 10
-    type: float
+    num: 5
+    var_type: float
+    type: jump
 
   gamma:
     min: 1e-5
     max: 1
+    num: 5
     warp: log
-    type: float
+    var_type: float
+    type: jump
 
 cv: 5
 
