### 2016-08-01 msmbuilder/osprey#163 @cxhernandez
jump var is parsed incorrectly
found this out while troubleshooting https://github.com/openjournals/joss-reviews/issues/34#issuecomment-236716258
