### 2016-08-01 1a11835
add new worker tests
