### 2016-08-02 #34 @cxhernandez
>```
Loading trials database: sqlite:////Users/Sebastian/Desktop/osprey-trials.db...
History contains: 10 trials
Choosing next hyperparameters with grid...
Error: GridSearchStrategy is defined only for all-enum search space
```

In order to use `GridSearchStrategy`, you would need to convert your search-space into enumerables using a `jump` variable:

```YAML
search_space:
  C:
    min: 0.1
    max: 10
    num: 10
    type: jump
    var_type: float

  gamma:
    min: 1e-5
    max: 1
    num: 10
    warp: log
    type: jump
    var_type: float
```

I realize this is not at all clear from our current documentation, however. Also, somewhat embarrassingly, I just now found a bug in how `jump` vars are parsed. Working on a [PR](https://github.com/msmbuilder/osprey/pull/164) now for a fix (it won't work otherwise) and additional docs + tests.

>When I choose
```YAML
strategy:
  name: gp
    return [var.point_to_gp(point_dict[var.name]) for var in self]
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/search_space.py", line 112, in <listcomp>
    return [var.point_to_gp(point_dict[var.name]) for var in self]
KeyError: 'C'
```

Not sure what's going on with `gp`. I can't reproduce this error locally. I'll add it as a unit test as well.

Thanks again for all your testing, @rasbt!
