### 2016-08-02 df46b1c
add random_state
