+++ osprey/data/grid_example.yaml
@@ -29,3 +29,5 @@ dataset_loader:
 
 trials:
     uri: sqlite:///osprey-trials.db
+
+random_seed: 42
