+++ osprey/data/random_example.yaml
@@ -25,3 +25,5 @@ dataset_loader:
 
 trials:
     uri: sqlite:///osprey-trials.db
+
+random_seed: 42
