### 2016-08-03 #34 @rasbt
> In order to use GridSearchStrategy, you would need to convert your search-space into enumerables using a jump variable

Oh yeah, that makes total sense, how would it now the increment space otherwise (unless it is hard coded). Thanks! (but adding it to the doc would be useful like you said).

> Also, somewhat embarrassingly, I just now found a bug in how jump vars are parsed.

Ah, sorry to hear that this review causes so much extra work, but I am happy to hear that it's been productive so far and that you discovered that bug :). And please take your time with the fixes; I think it would be worthwhile bumping the version to e.g., 1.1 and also updating the documentation in one swoop then before we get back to the review.
