### 2016-08-03 #34 @cxhernandez
> I think it would be worthwhile bumping the version to e.g., 1.1 and also updating the documentation in one swoop then before we get back to the review.

Agreed! I'll let you know when it's ready to be tested (shouldn't take too long, hopefully). Thanks again!
