### 2016-08-05 f77979a
update license info
