+++ README.md
@@ -3,7 +3,7 @@ Osprey
 [![Build Status](https://travis-ci.org/msmbuilder/osprey.svg?branch=master)](https://travis-ci.org/msmbuilder/osprey)
 [![Coverage Status](https://coveralls.io/repos/github/msmbuilder/osprey/badge.svg?branch=master)](https://coveralls.io/github/msmbuilder/osprey?branch=master)
 [![PyPi version](https://badge.fury.io/py/osprey.svg)](https://pypi.python.org/pypi/osprey/)
-[![License](https://img.shields.io/badge/license-ASLv2.0-red.svg?style=flat)]  (https://pypi.python.org/pypi/osprey/)
+[![License](https://img.shields.io/badge/license-ASLv2.0-red.svg?style=flat)]  (http://www.apache.org/licenses/LICENSE-2.0)
 [![DOI](https://zenodo.org/badge/9890/msmbuilder/osprey.svg)](https://zenodo.org/badge/latestdoi/9890/msmbuilder/osprey)
 [![Documentation](https://img.shields.io/badge/docs-latest-blue.svg?style=flat)] (http://msmbuilder.org/osprey)
 
