### 2016-08-05 #34 @cxhernandez
@arfon, @mpharrigan, @rasbt

I figured out a hack for the affiliations. If you add these lines to the template:

```latex
$if(authors)$
  $for(authors)$
  $if(authors.affiliation)$
    \author[$authors.affiliation$]{$authors.name$}
  $else$
    \author{$authors.name$}
  $endif$
  $endfor$
$endif$

$if(affiliations)$
  $for(affiliations)$
    \affil[1]{$affiliations$}
  $endfor$
$endif$
```

and then have this in the `paper.md` header:

```markdown
authors:
  - name: Robert T. McGibbon
    affiliation: 1
  - name: Carlos X. Hernández
    orcid: 0000-0002-8146-5904
    affiliation: 1
  - name: Matthew P. Harrigan
    affiliation: 1
  - name: Steven Kearnes
    affiliation: 1
  - name: Mohammad M. Sultan
    affiliation: 1
  - name: Stanislaw Jastrzebski
    affiliation: 2
  - name: Brooke E. Husic
    affiliation: 1
  - name: Vijay S. Pande
    affiliation: 1
affiliations:
  - Stanford University
  - Jagiellonian University
```

You get the desired result:

![image](https://cloud.githubusercontent.com/assets/5142365/17426387/11a26892-5a8c-11e6-90d6-9485ead4451f.png)
