### 2016-08-05 #34 @rasbt
Nice! I think it would be better to have the correct indices on the affiliations as well (currently both "1"s in the example above)

Here, you need to replace the `[1]` in

```
$if(affiliations)$
  $for(affiliations)$
    \affil[1]{$affiliations$}
  $endfor$
$endif$
```
by a counter variable `1, ..., n_affiliations`
