### 2016-08-05 #34 @cxhernandez
Ah! I jumped the gun! Can you even use counters in pandoc templates? I'm googling, but nothing's coming up.
