### 2016-08-05 #34 @rasbt
I think you can do something like this:

```
affiliations:
  - {name: Stanford University, index: 1}
  - {name: Jagiellonian University, index:1}
```

and then

```
$if(affiliations)$
  $for(affiliations)$
    \affil[affiliations.index]{$affiliations.name$}
  $endfor$
$endif$
```
