### 2016-08-05 04facd6
update installation docs
