+++ requirements.txt
@@ -1,0 +1,8 @@
+six>=1.10.0
+pyyaml>=3.11
+numpy>=1.10.4
+scipy>=0.17.0
+scikit-learn>=0.17.0
+sqlalchemy>=1.0.10
+bokeh>=0.12.0
+matplotlib>=1.5.0
