### 2016-08-05 1fa3daf
add contributing section
