+++ docs/index.rst
@@ -5,10 +5,10 @@ Osprey is a tool for practical hyperparameter optimization of machine learning
 algorithms. It's designed to provide a practical, **easy to use** way for
 application scientists to find parameters that maximize the cross-validation
 score of a model on their dataset. Osprey is actively being developed by
-`researchers at Stanford University <https://pande.stanford.edu>`_, with primary
-application areas in computational protein dynamics and drug design,
-and distributed under the
+researchers around the world, with primary application areas in computational
+protein dynamics and drug design, and distributed under the
 `Apache License (v2.0) <https://www.apache.org/licenses/LICENSE-2.0>`_.
+All development takes place on `GitHub <https://github.com/msmbuilder/osprey>`_.
 
 Overview
 --------
@@ -27,6 +27,10 @@ Related tools include and `spearmint <https://github.com/JasperSnoek/spearmint>`
 
 To get started, run ``osprey skeleton`` to create an example config file, and
 then boot up one or more parallel instances of ``osprey worker``.
+
+If you happen to run into any issues while using Osprey or would like suggest a
+new feature, please take a moment to read our :ref:`Contributing <contributing>`
+section.
 
 
 .. raw:: html
