### 2016-08-05 5b9423a
update config docs
