### 2016-08-05 msmbuilder/osprey#163 @cxhernandez
addressed in https://github.com/msmbuilder/osprey/pull/164
