### 2016-08-05 06e9974
Update README.md
be consistent with capitalization
