### 2016-08-05 #34 @cxhernandez
Alright, I've incorporated the majority of the suggestions listed above into the `master` branch. I'm holding off on making a new release (v1.1.0) until this review is complete (in case further changes are necessary).

> Also, I am curious how I can run osprey on my own dataset. What format should the file be in? Say I have an iris.csv file from the UCI repo, how would I feed it to osprey?

Sorry @rasbt, I must have missed this question before. At the moment, `.csv` is not supported in our current stable of `dataset_loaders` (although it should be trivial to add); however, we currently support `.npy`, `.hdf5`, `joblib` files, and pretty much every major MD trajectory format (through [`mdtraj`](https://github.com/mdtraj/mdtraj)). For the iris dataset, you might find the [`SklearnDatasetLoader `](https://github.com/msmbuilder/osprey/blob/master/osprey/dataset_loaders.py#L168) handy:

```YAML
dataset_loader:
  name: sklearn_dataset
  params:
    method: load_iris
```
