### 2016-08-05 #34 @rasbt
Sounds great, I will try to give it a look this weekend then!

> however, we currently support .npy, hdf5, joblib

Still, I am wondering how the input format looks like? E.g., for classifiers, is it samples per row and features per column, and the class label in the last column (or first column, or separate object?)? I.e., let's say that I am a new user and I want to run osprey on my own non-" MD trajectory" dataset (as far as I understand from the paper, it is meant to be a general tool, not protein chemistry specific tool?). How do I do it; I think a little bit more guidance would be required.

>  For the iris dataset, you might find the SklearnDatasetLoader handy:

Sure, but I asked because I was curious how to prepare a dataset to run it in osprey. Let's pretend it's my private research dataset. How would I get it into osprey?
