### 2016-08-05 #34 @cxhernandez
> Still, I am wondering how the input format looks like?

The way this works at the moment is similar to `sklearn.datasets.load_files`. You would break up your dataset into many different files, each file representing a unique class label (eg. 'class_1.h5', ..., 'class_n. h5'). The datasets should be organized so that each column represents a feature and each row is an observation. I can add more documentation on this.
