### 2016-08-05 #34 @mpharrigan
It would probably be helpful to add a tutorial/example using a non-biophysics dataset
