### 2016-08-06 #34 @rasbt
It would be nice @cxhernandez if you could add the instructions for preparing a CSV file for osprey. Or maybe not necessarily a CSV file, but a NumPy data array of shape [n_rows, n_samples] and for the supervised methods a target variable array. Sorry for all the extra work, but In my opinion, this would be necessary to make it available to a general audience and to pass the JOSS requirements, such as

>  Example usage: Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).
>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g. API method documentation)?
