### 2016-08-07 msmbuilder/osprey#173 @cxhernandez
Delimiter-separated dataset loader 
Would be useful to broaden our user-base (https://github.com/openjournals/joss-reviews/issues/34#issuecomment-237858608)
