### 2016-08-07 msmbuilder/osprey#174 @cxhernandez
Add docs on how to prepare a dataset for Osprey
Currently not very clear how to load data for non-MD datasets (https://github.com/openjournals/joss-reviews/issues/34#issuecomment-238003812)
