### 2016-08-08 a6d55b1
Add "module" estimator spec
Import a (possibly local) module and look for a global named `estimator`
