### 2016-08-08 msmbuilder/osprey#173 @cxhernandez
done in #175
