### 2016-08-11 c562b20
added data formatting docs
