### 2016-08-15 msmbuilder/osprey#174 @cxhernandez
done in #185
