+++ docs/getting_started.rst
@@ -1,4 +1,3 @@
-
 Getting Started
 ===============
 
@@ -13,10 +12,10 @@ dataset information, cross-validation strategy, and a path to a
 generate an example configuration file.
 
 First, we will describe how to prepare your dataset for Osprey. Then, we will
-show how to use Osprey for a simple scikit-learn classification task. And
-finally, we will show how one might use Osprey to model a
+show how to use Osprey for a simple scikit-learn classification task. We'll
+also demonstrate how one might use Osprey to model a
 `molecular dynamics (MD) <https://en.wikipedia.org/wiki/Molecular_dynamics>`_
-dataset.
+dataset. And finally, we'll show how to query and use your final Osprey results.
 
 
 Formatting Your Dataset
@@ -102,7 +101,7 @@ we write:
       params:
         method: load_digits
 
-And finally we need to list the SQL database where our cross-validation
+And finally we need to list the SQLite database where our cross-validation
 results will be saved:
 
 .. code:: yaml
@@ -210,7 +209,7 @@ files as a glob-string:
         topology: ~/local/msmbuilder/Tutorial/native.pdb
         stride: 1
 
-And finally we need to list the SQL database where our cross-validation
+And finally we need to list the SQLite database where our cross-validation
 results will be saved:
 
 .. code:: yaml
@@ -225,3 +224,115 @@ we can start an osprey job in the command-line by invoking:
 .. code:: bash
 
     $ osprey worker config.yaml
+
+
+Working with Osprey Results
+---------------------------
+
+As mentioned before, all Osprey results are stored in an SQL-like database, as
+define by the ``trials`` field in the configuration file. This makes querying
+and reproducing Osprey results fairly simple.
+
+Osprey provides two command-line tools to quickly digest your results:
+``current_best`` and ``plot``. ``current_best``, as the name suggests, prints
+out the best scoring model currently in your trials database, as well as the
+parameters used to create it. Here's some example output from our SVM
+classification tutorial above:
+
+.. code:: bash
+
+    $ osprey current_best config.yaml
+
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    Best Current Model = 0.975515 +- 0.013327
+    ~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~
+    SVC(C=1.0, cache_size=200, class_weight=None, coef0=0.0,
+      decision_function_shape=None, degree=3, gamma='auto', kernel='rbf',
+      max_iter=-1, probability=False, random_state=None, shrinking=True,
+      tol=0.001, verbose=False)
+    		 C 	 7.957695018309156
+    		 gamma 	 0.0004726222555749291
+
+
+This is useful if you just want to get a sense of how well your trials are
+doing or just want to quickly get the best current result from Osprey.
+The ``plot`` functionality provides interactive HTML charts using ``bokeh``
[…]
