### 2016-08-16 1371fbf
add trial_results
