+++ requirements.txt
@@ -6,3 +6,4 @@ scikit-learn>=0.17.0
 sqlalchemy>=1.0.10
 bokeh>=0.12.0
 matplotlib>=1.5.0
+pandas>=0.18.0
