### 2016-08-16 8b9f432
Update config_file.rst
