### 2016-08-16 d651f35
Update config_file.rst
