### 2016-08-16 5d409f8
Update index.rst
