### 2016-08-16 b33cdcb
Update getting_started.rst
