### 2016-08-16 #34 @cxhernandez
@rasbt: We've added more complete instructions [here](http://msmbuilder.org/osprey/development/getting_started.html)

Let me know what you think!
