### 2016-08-16 #34 @rasbt
Nice, I will try to give it a whirl over the next couple of days.
Thanks for adding the random_state param. Just skimming through the documentation, I read that

> In case you need reproducible Osprey trials, you can also include an optional random seed as seen below:
>
> Example:
> random_seed: 42
> Please note that this makes parallel trials redundant and, thus, not recommended when scaling across multiple jobs.

So that means that every parallel trial will be run with the same random state? I think a simple workaround would be to increment the random seed for each parallel run so that they are different (and useful) but still reproducible.
