### 2016-08-16 53bdf3e
Update getting_started.rst
