### 2016-08-16 #34 @cxhernandez
> I think a simple workaround would be to increment the random seed for each parallel run so that they are different (and useful) but still reproducible.

Exactly. One would have to create a separate configuration file for each parallel worker in this case; I can add a note about this.
