### 2016-08-16 0f55414
Update config_file.rst
add note re: https://github.com/openjournals/joss-reviews/issues/34#issuecomment-240151698
