### 2016-08-30 99c0e16
add seed to CLI
