### 2016-08-30 #34 @cxhernandez
@rasbt: I've added the ability to specify the random seed through the CLI, which should make it much more convenient to generate many reproducible workers
