### 2016-08-31 #34 @rasbt
Hey, just got the latest release from GitHub and were trying to run the unit tests via `nose` and pumped into another little problem

```
    return _load(spec)
  File "<frozen importlib._bootstrap>", line 693, in _load
  File "<frozen importlib._bootstrap>", line 673, in _load_unlocked
  File "<frozen importlib._bootstrap_external>", line 665, in exec_module
  File "<frozen importlib._bootstrap>", line 222, in _call_with_frames_removed
  File "/Users/Sebastian/Desktop/osprey-master/osprey/__init__.py", line 2, in <module>
    from .version import version as _version
ImportError: No module named 'osprey.version'

Ran 1 test in 0.015s

FAILED (errors=1)
(osprey) ~/Desktop/osprey-master$
```

Now, I am wondering, which Python version(s) does osprey support? 2.7, 3.5, others? I would recommend adding a note about that in the README and on the documentation page itself in the Installation section
