### 2016-08-31 #34 @rasbt
Another little problem I bumped into when I was trying to run Osprey on a CSV version of Iris; hope you could help with that :)

I've setup the parameter file as follows

```
estimator:
  entry_point: sklearn.linear_model.LogisticRegression

strategy:
  name: grid

search_space:
  logistic__C:
    min: 1e-3
    max: 1e3
    type: float
    warp: log
  logistic__penalty:
    choices:
      - l1
      - l2
    type: enum

dataset_loader:
  name: dsv
  params:
    filenames: /Users/Sebastian/Desktop/iris.csv
    delimiter: ','
    y_col: 4
    usecols: 0, 1, 2, 3

cv: 5

trials:
    uri: sqlite:///osprey-trials.db

random_seed: 42
```

I also tried adding

```
    skip_header: 0
    skip_footer: 0
```

but got the same error message. Also, I tried `usecols: 0, 1, 2, 3, 4` instead of `usecols: 0, 1, 2, 3` but it didn't, same error message:

```
Traceback (most recent call last):
  File "/Users/Sebastian/miniconda3/envs/osprey/bin/osprey", line 9, in <module>
    load_entry_point('osprey==1.1.0.dev0', 'console_scripts', 'osprey')()
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/cli/main.py", line 37, in main
    args_func(args, p)
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/cli/main.py", line 42, in args_func
    args.func(args, p)
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/cli/parser_worker.py", line 8, in func
    execute(args, parser)
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/execute_worker.py", line 45, in execute
    X, y = config.dataset()
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/config.py", line 284, in dataset
    X, y = loader.load()
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/dataset_loaders.py", line 146, in load
    data = self.transform(self.loader(fn))
  File "/Users/Sebastian/miniconda3/envs/osprey/lib/python3.5/site-packages/osprey/dataset_loaders.py", line 131, in transform
    return X[::self.stride, x_idx], X[::self.stride, y_idx].ravel()
IndexError: index 1 is out of bounds for axis 1 with size 1
```

The iris.csv is formatted as follows:

```
5.1,3.5,1.4,0.2,0
4.9,3,1.4,0.2,0
4.7,3.2,1.3,0.2,0
4.6,3.1,1.5,0.2,0
5,3.6,1.4,0.2,0
5.4,3.9,1.7,0.4,0
4.6,3.4,1.4,0.3,0
5,3.4,1.5,0.2,0
...
```
(sepal length, sepal width, petal length, petal width, class label)

I also attached the iris.csv for your reference as a zip since GitHub doesn't let me upload CSVs for some reason

[iris.csv.zip](https://github.com/openjournals/joss-reviews/files/446289/iris.csv.zip)
