### 2016-08-31 #34 @cxhernandez
@rasbt: Did you share the correct csv file?

```bash
(msmb-dev) [cxh@local ~/Developer/osprey/example]$ cat iris.csv
5.9,3,5.1,1.8,2
```

It seems to only have one row (which `np.genfromtxt` interprets as a 1-D array).

```python
In [8]: np.genfromtxt(self.filenames, delimiter=self.delimiter,
   ...:               skip_header=self.skip_header,
   ...:               skip_footer=self.skip_footer,
   ...:               filling_values=self.filling_values,
   ...:               usecols=self.usecols)
Out[8]: array([ 5.1,  3.5,  1.4,  0.2,  nan])
```
