### 2016-08-31 #34 @cxhernandez
This error seems to stem from how `numpy.genfromtxt` handles the DOS carriage return characters in your CSV file. I [replaced them](https://github.com/openjournals/joss-reviews/files/447830/iris.unix.txt) with a newline character and the data-loading seems to work just as intended. I'm inclined to report this to the NumPy folks, as this shouldn't be a difficult fix.

As far as the configuration file goes, the hyperparameters should be defined as simply `C` and `penalty` as the estimator is not part of a pipeline where `LogisticRegression` is defined as `'logistic'`. Also, since you're using the grid search strategy, `C` should be defined as a [jump variable](http://msmbuilder.org/osprey/development/config_file.html#search-space) to convert the float into an enumerable:

```YAML
  C:
    min: 1e-3
    max: 1e3
    num: 10
    type: jump
    var_type: float
    warp: log
```
