### 2016-08-31 #34 @cxhernandez
> Now, I am wondering, which Python version(s) does osprey support? 2.7, 3.5, others? I would recommend adding a note about that in the README and on the documentation page itself in the Installation section

I can add this to the README and docs . Osprey should support Python 2.7+.

As far as that particular error goes, you need to run `python setup.py install` (or `develop`) prior to nose testing so that the `version` module can be generated. Might be worth adding a `testing` section to the docs as well.
