### 2016-08-31 #34 @rasbt
> This error seems to stem from how numpy.genfromtxt handles the DOS carriage return characters in your CSV file.

Ah, you are right, I exported the Iris feature array via numpy and then noticed that I forgot the class labels, which I quickly added via Excel. Excel must have replaced the newline chars by these weird  DOS carriage return characters. Looked normal to me in my text editor (atom), but I had the same 1-line problem when using `cat`. Thanks for the hint!

> As far as the configuration file goes, the hyperparameters should be defined as simply C and penalty as the estimator is not part of a pipeline where LogisticRegression is defined as 'logistic'. Also, since you're using the grid search strategy, C should be defined as a jump variable to convert the float into an enumerable:

Sorry about that, I remember you mentioned that earlier, and I just didn't pay attention when I copy&pasted the hyperparam settings from the documentations and changed it to `grid`
