### 2016-08-31 #34 @rasbt
Works, nicely now! :).

Have one more question about the random state though. Where exactly is it used? Only in the cross-validator? Because I noticed that the `random_state` is set to "null" in the estimator itself

![screen shot 2016-08-31 at 1 21 02 pm](https://cloud.githubusercontent.com/assets/5618407/18139107/f51e5d8c-6f7d-11e6-958a-19b8fe4e110b.png)

The param file I was using looks as follows (where I set the random state at the bottom)

```
stimator:
  entry_point: sklearn.linear_model.LogisticRegression

strategy:
  name: grid

search_space:
  solver:
    choices:
      - newton-cg
    type: enum
  multi_class:
    choices:
      - multinomial
    type: enum
  C:
    min: 1e-3
    max: 1e3
    num: 10
    type: jump
    var_type: float
    warp: log
  penalty:
    choices:
      - l2
    type: enum

dataset_loader:
  name: dsv
  params:
    filenames: /Users/Sebastian/Desktop/iris-2.csv
    delimiter: ','
    y_col: 4
    usecols: 0, 1, 2, 3

cv: 5

trials:
    uri: sqlite:///osprey-trials.db

random_seed: 42
```

However, adding

```
  random_state:
    choices:
      - 123
    type: enum
```

seems to do exactly what I want, so it's not necessary to fix something, I am just curious :)
