### 2016-08-31 #34 @rasbt
Ah, I am having troubles with nosetests again :P
So, I ran `python setup.py develop` and it solved the `version` issue, another thing though is that somehow nosetests is not grabbing the tests.

```
~/Desktop/osprey-master$ nosetests

Ran 0 tests in 0.031s

OK
~/Desktop/osprey-master$ ls
LICENSE			basesetup.py		osprey			requirements.txt
README.md		devtools		osprey.egg-info		setup.py
__pycache__		docs			paper
~/Desktop/osprey-master$ ls osprey/tests/
__init__.py			test_config.py			test_search_space.py
__pycache__			test_dataset_loader.py		test_strategies.py
test_cli_skeleton.py		test_entry_point.py		test_trials.py
test_cli_worker_and_dump.py	test_fit_estimator.py		test_utils.py
~/Desktop/osprey-master$ cd ~/code/siteinterlock
~/code/siteinterlock$ nosetests
....................
Ran 20 tests in 5.663s

OK
```

Maybe I am doing something silly or overlooking something ... :/
