### 2016-08-31 #34 @cxhernandez
> Have one more question about the random state though. Where exactly is it used? Only in the cross-validator? Because I noticed that the random_state is set to "null" in the estimator itself

It's used [here](https://github.com/msmbuilder/osprey/blob/master/osprey/fit_estimator.py#L43) to seed NumPy's random generator But you make a good point that this should also be passed to the estimator explicitly.
