### 2016-08-31 #34 @jkahn
Ok I'm self-appointing as the [review-thread police](https://github.com/openjournals/joss/issues/163) -- can @cxhernandez and @rasbt carry on this debugging on a separate issue filed on [the osprey issue tracker](https://github.com/msmbuilder/osprey/issues) please?
