### 2016-08-31 msmbuilder/osprey#197 @cxhernandez
Pass random_seed to estimator
re: https://github.com/openjournals/joss-reviews/issues/34#issuecomment-243837870

In practice, this shouldn't make a difference, but it's nice to have a record of the random seed used.
