### 2016-08-31 #34 @rasbt
Okay, I won't go much into debugging then, I think it's fine its current state (these are just points you may want to add in future versions). I checked off all the points, and the only thing I would do then for version 1.1.0 is to add the required Python versions in the dependency list (e.g., mentioning that it works for both Python 2.7.x and 3.5)

## Software

- [x] tests

There's currently only a 65% coverage; I am not sure what the JOSS requirements are, personally I would recommend extending the unit tests to at least 90%.

- [x] License

## The JOSS paper

The JOSS paper (the PDF associated with this submission) should only include:

- [x] A list of the authors of the software

- [x] Author affiliations

- [x] A short summary describing the high-level functionality of the software

- [x] A list of key references including a link to the software archive

- [x] A high-level overview of this documentation should be included in a README file (or equivalent).

## A statement of need

- [x] The authors should clearly state what problems the software is designed to solve and who the target audience is.

## Installation instructions

- [ ] There should be a clearly-stated list of dependencies. Ideally these should be handled with an automated package management solution.

Adding the required/supported Python version would be great.

## Example usage

- [x] The authors should include examples of how to use the software (ideally to solve real-world analysis problems).

## API documentation

- [x] Reviewers should check that the software API is documented to a suitable level.

- [x] Authors are strongly encouraged to include an automated test suite covering the core functionality of their software.

There should be clear guidelines for third-parties wishing to:

- [x] Contribute to the software

- [x] Report issues or problems with the software

- [x] Seek support

> Include here some examples of well-documented software for people to review.

E.g., scikit-learn, Keras, or TensorFlow

- [x] Reviewers are expected to install the software they are reviewing and to verify the core functionality of the software.

## Other Notes

- [ ] Once everything is set, make new stable release
- [ ] change version label accordingly
