### 2016-09-01 a470329
pass random_seed to estimator
