+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,4 +1,5 @@
  - [x] Implement feature / fix bug
  - [ ] Add tests
+ - [ ] Update changelog
 
 [Describe changes here]
