### 2016-09-01 msmbuilder/osprey#197 @cxhernandez
done in https://github.com/msmbuilder/osprey/pull/198
