+++ paper/paper.bib
@@ -66,7 +66,7 @@ url="http://dx.doi.org/10.1023/A:1008306431147"
                   Stanislaw Jastrzebski and
                   Brooke E. Husic and
                   Vijay Pande},
-  title        = {osprey: Osprey 1.0.0},
+  title        = {Osprey 1.0.0},
   month        = jun,
   year         = {2016},
   doi          = {10.5281/zenodo.56251},
