### 2016-09-01 #34 @cxhernandez
@arfon, @rasbt: I've added the most recent [suggestions](https://github.com/openjournals/joss-reviews/issues/34#issuecomment-243890240) and have [v1.1.0](https://github.com/msmbuilder/osprey/pull/199) ready to be merged, pending anything else.
