### 2016-09-04 #34 @arfon
>  I would suggest changing the version tag (to 1.1.0) in the paper as well though, since this will be the one uploaded to JOSS if I understand correctly.

Will do. So are we good to accept @rasbt?
