### 2016-09-05 #34 @rasbt
> Will do. So are we good to accept @rasbt?

Yes, it looks good to me :).
