### 2016-09-08 #34 @arfon
Thanks for the review @rasbt!

@cxhernandez your paper is now accepted into JOSS and your DOI is http://dx.doi.org/10.21105/joss.00034 🚀 🎉 💥
