### 2016-09-08 #34 @cxhernandez
The software (and docs) undeniably benefitted from the review process! I look forward to submitting more stuff 😄

Thanks again!
